+++ src/main/java/de/dm/intellij/liferay/index/ActionCommandIndex.java
@@ -98,8 +98,8 @@ public class ActionCommandIndex extends FileBasedIndexExtension<CommandKey, Void
         return true;
     }
 
-    public static List<String> getActionCommands(@NotNull String portletName, GlobalSearchScope scope) {
-        return AbstractCommandKeyIndexer.getCommands(NAME, portletName, scope);
+    public static List<String> getActionCommands(@NotNull String portletName, Project project, GlobalSearchScope scope) {
+        return AbstractCommandKeyIndexer.getCommands(NAME, portletName, project, scope);
     }
 
     public static List<PsiFile> getPortletClasses(Project project, String portletName, String commandName, GlobalSearchScope scope) {