+++ src/main/java/de/dm/intellij/liferay/index/PortletJspIndex.java
@@ -6,37 +6,23 @@ import com.intellij.openapi.application.ReadAction;
 import com.intellij.openapi.project.IndexNotReadyException;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.text.StringUtil;
-import com.intellij.psi.PsiClass;
-import com.intellij.psi.PsiExpression;
-import com.intellij.psi.PsiLiteralExpression;
-import com.intellij.psi.PsiMethod;
-import com.intellij.psi.PsiModifierList;
-import com.intellij.psi.PsiReturnStatement;
+import com.intellij.psi.*;
 import com.intellij.psi.search.GlobalSearchScope;
 import com.intellij.psi.util.PsiUtil;
-import com.intellij.util.indexing.DataIndexer;
-import com.intellij.util.indexing.DefaultFileTypeSpecificInputFilter;
-import com.intellij.util.indexing.FileBasedIndex;
-import com.intellij.util.indexing.FileBasedIndexExtension;
-import com.intellij.util.indexing.FileContent;
-import com.intellij.util.indexing.ID;
-import com.intellij.util.indexing.PsiDependentIndex;
+import com.intellij.util.indexing.*;
 import com.intellij.util.io.DataExternalizer;
 import com.intellij.util.io.EnumeratorStringDescriptor;
 import com.intellij.util.io.KeyDescriptor;
 import com.intellij.util.io.VoidDataExternalizer;
 import de.dm.intellij.liferay.util.LiferayFileUtil;
+import de.dm.intellij.liferay.util.ProjectUtils;
 import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 import static de.dm.intellij.liferay.util.ProjectUtils.getMethodParameterQualifiedNames;
 
@@ -135,7 +121,7 @@ public class PortletJspIndex extends FileBasedIndexExtension<JspKey, Void> imple
                         NAME,
                         jspKey -> {
                             if (jspPath.equals(jspKey.getJspPath())) {
-                                result.add(PortletNameIndex.resolvePortletName(jspKey.getPortletName(), project, scope));
+                                result.add(ProjectUtils.resolveReferencePlaceholder(jspKey.getPortletName(), project, scope));
                             }
                             return true;
                         },
@@ -176,7 +162,7 @@ public class PortletJspIndex extends FileBasedIndexExtension<JspKey, Void> imple
                         for (String initParamValue : initParamValues) {
                             for (String portletName : portletNames) {
                                 String portletId = portletName;
-                                if (! portletName.startsWith(AbstractComponentPropertyIndexer.REFERENCE_PLACEHOLDER)) {
+                                if (! portletName.startsWith(ProjectUtils.REFERENCE_PLACEHOLDER)) {
                                     portletId = LiferayFileUtil.getPortletId(portletName);
                                 }