+++ src/main/java/de/dm/intellij/liferay/index/ResourceCommandIndex.java
@@ -89,8 +89,8 @@ public class ResourceCommandIndex extends FileBasedIndexExtension<CommandKey, Vo
         return true;
     }
 
-    public static List<String> getResourceCommands(@NotNull String portletName, GlobalSearchScope scope) {
-        return AbstractCommandKeyIndexer.getCommands(NAME, portletName, scope);
+    public static List<String> getResourceCommands(@NotNull String portletName, Project project, GlobalSearchScope scope) {
+        return AbstractCommandKeyIndexer.getCommands(NAME, portletName, project, scope);
     }
 
     public static List<PsiFile> getPortletClasses(Project project, String portletName, String commandName, GlobalSearchScope scope) {