+++ testdata/de/dm/intellij/liferay/index/ActionCommandIndexTest/de/dm/action/ActionKeys.java
@@ -1,0 +1,8 @@
+package de.dm.action;
+
+public class ActionKeys {
+
+    public static final String MY_ACTION_KEY = "/my/actionkey";
+    public static final String MY_SECOND_ACTION_KEY = "/my/secondkey";
+
+}