+++ testdata/de/dm/intellij/liferay/index/ActionCommandIndexTest/de/dm/action/MyActionKeyReferenceMVCAction.java
@@ -1,0 +1,17 @@
+package de.dm.action;
+
+import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
+import org.osgi.service.component.annotations.Component;
+import de.dm.action.ActionKeys;
+
+@Component(
+    immediate = true,
+    property = {
+        "javax.portlet.name=de_dm_portlet_MyPortletName",
+        "mvc.command.name=" + ActionKeys.MY_ACTION_KEY
+    },
+    service = MVCActionCommand.class
+)
+public class MyReferenceMVCAction implements MVCActionCommand {
+
+}