+++ testdata/de/dm/intellij/liferay/index/ActionCommandIndexTest/de/dm/portlet/PortletKeys.java
@@ -3,6 +3,6 @@ package de.dm.portlet;
 public class PortletKeys {
 
     public static final String MY_PORTLET_KEY = "de_dm_portlet_MyPortletKey";
-    public static final String MY_REFERENCE_PORTLET_KEY = "de_dm_portlet_MyReferencePortletKey";
+    public static final String MY_SECOND_PORTLET_KEY = "de_dm_portlet_MySecondPortletKey";
 
 }