+++ testdata/de/dm/intellij/liferay/language/jsp/LiferayTaglibActionCommandNameReferenceContributorTest/de/dm/action/ActionKeys.java
@@ -1,0 +1,7 @@
+package de.dm.action;
+
+public class ActionKeys {
+
+    public static final String MY_ACTION_KEY = "/my/actionkey";
+
+}